integrate sandbox-go command to azure functions
