+++ cmd/azurefunctions/azurefunctions.go
@@ -19,7 +19,7 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 */
-package cmd
+package azurefunctions
 
 import (
 	"fmt"
@@ -27,31 +27,16 @@ import (
 	"github.com/spf13/cobra"
 )
 
-// httpexampleCmd represents the httpexample command
-var httpexampleCmd = &cobra.Command{
-	Use:   "httpexample",
-	Short: "A brief description of your command",
-	Long: `A longer description that spans multiple lines and likely contains examples
-and usage of using your command. For example:
-
-Cobra is a CLI library for Go that empowers applications.
-This application is a tool to generate the needed files
-to quickly create a Cobra application.`,
+// azurefunctionsCmd represents the azurefunctions command
+var azurefunctionsCmd = &cobra.Command{
+	Use:   "azurefunctions",
+	Short: "Azure Functions Command",
+	Long:  `This is a azurefunctions command.`,
 	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println("httpexample called")
+		fmt.Println("azurefunctions called")
 	},
 }
 
-func init() {
-	azurefunctionsCmd.AddCommand(httpexampleCmd)
-
-	// Here you will define your flags and configuration settings.
-
-	// Cobra supports Persistent Flags which will work for this command
-	// and all subcommands, e.g.:
-	// httpexampleCmd.PersistentFlags().String("foo", "", "A help for foo")
-
-	// Cobra supports local flags which will only run when this command
-	// is called directly, e.g.:
-	// httpexampleCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+func GetCommand() *cobra.Command {
+	return azurefunctionsCmd
 }
