+++ scripts/deploy_azure_functions_resources.sh
@@ -30,12 +30,3 @@ az functionapp create \
     --name "$FUNCTION_APP_NAME" \
     --os-type linux \
     --storage-account "$STORAGE_NAME"
-
-# # create json file containing the resource group name and storage account name
-# echo "{\"RESOURCE_GROUP_NAME\": \"$RESOURCE_GROUP_NAME\", \"STORAGE_NAME\": \"$STORAGE_NAME\", \"FUNCTION_APP_NAME\": \"$FUNCTION_APP_NAME\"}" > azure-functions.json
-
-# # Build the function app
-# GOOS=linux GOARCH=amd64 CGO_ENABLED=0 go build -ldflags='-s -w' -trimpath -o dist/func ./HttpExample/main.go
-
-# # Deploy the function app
-# func azure functionapp publish "$FUNCTION_APP_NAME"
